+++ src/control_pkg/include/Imu.h
@@ -15,12 +15,12 @@ public:
     ~Imu();
     void Imu_Initial();
     void Imu_Update();
-    Eigen::Matrix<float,3,3> GetRotMat();
-    Eigen::Matrix<float,3,1> GetAcc();
-    Eigen::Matrix<float,3,1> GetGyro();
-    Eigen::Matrix<float,4,1> GetQuat();
-    Eigen::Matrix<float,3,1> getAccGlobal();
-    Eigen::Matrix<float,3,1> getGyroGlobal();
+    Eigen::Matrix<double,3,3> GetRotMat();
+    Eigen::Matrix<double,3,1> GetAcc();
+    Eigen::Matrix<double,3,1> GetGyro();
+    Eigen::Matrix<double,4,1> GetQuat();
+    Eigen::Matrix<double,3,1> getAccGlobal();
+    Eigen::Matrix<double,3,1> getGyroGlobal();
     float getPitch();
     float getRoll();
     float getYaw();
