+++ src/control_pkg/include/Reversal_solution.h
@@ -7,12 +7,12 @@
 #include "mathtool.h"
 #include "Kenimatics_normal_solution.h"
 
-inline float clamp(float val, float min_val, float max_val);
-float q1_ik(float py, float pz, float l1);
-float q3_ik(float b3z, float b4z, float b);
-float q2_ik(float q1, float q3, float px, float py, float pz, float b3z, float b4z);
-Eigen::Matrix<float,3,1> Reversal_Solution_Update(uint8_t group , float x , float y , float z);
-Eigen::Matrix<float,3,1> Reversal_Update_B(uint8_t group , float x , float y , float z);
+inline double clamp(double val, double min_val, double max_val);
+double q1_ik(double py, double pz, double l1);
+double q3_ik(double b3z, double b4z, double b);
+double q2_ik(double q1, double q3, double px, double py, double pz, double b3z, double b4z);
+Eigen::Matrix<double,3,1> Reversal_Solution_Update(uint8_t group , double x , double y , double z);
+Eigen::Matrix<double,3,1> Reversal_Update_B(uint8_t group , double x , double y , double z);
 Vec12 Reversal_GetQ(const Vec34 &vecP, FrameType frame);
 // Eigen::Matrix<double,3,3> calcJaco(int legid , Eigen::Matrix<float,3,1> q);
 Vec3 calcQd(int legid,Vec3 pEe, Vec3 vEe, FrameType frame);
